+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 * The new version of the [web3js-eea library (v0.10)](https://github.com/PegaSysEng/web3js-eea) supports the onchain privacy group management changes made in Besu v1.5.3.     
 
 ### Bug Fixes
-* Fix logs queries missing results against chain head [\#1351](https://github.com/hyperledger/besu/pull/1351)
+* Added `debug_getBadBlocks` JSON-RPC API to analyze and detect consensus flaws. Even if a block is rejected it will be returned by this method [\#1378](https://github.com/hyperledger/besu/pull/1378)
+* Fix logs queries missing results against chain head [\#1351](https://github.com/hyperledger/besu/pull/1351) and [\#1381](https://github.com/hyperledger/besu/pull/1381) 
 
 #### Previously identified known issues